Make default_char an attribute
screen.default_char is not longer a property but an attribute. This
avoids calling over an over a function to compute the default char
based on the screen.mode when most of the time this mode does not
change.

Instead, screen.default_char remains constant until a change in the mode
happen where update_default_char() is called to update the attribute
default_char.

Code that reads the default_char will not longer have to pay the cost of
a function call (just an attribute lookup).
Code that needs to mutate the default_char they need to get a fresh
copy calling new_empty_char()